+++ packages/react-web3-icons/src/components/UnknownIcon.tsx
@@ -1,0 +1,10 @@
+import React, { ComponentProps } from "react";
+
+import { iconUnknown } from "../icons/full/build/icon-unknown.icon";
+import { Image } from "./Image";
+
+type UnknownIconProps = ComponentProps<"img">;
+
+export const UnknownIcon = (props: UnknownIconProps) => {
+  return <Image svgCode={iconUnknown.data} {...props} />;
+};
