+++ packages/react-web3-icons/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./components/index";
+export { formatMonoSvgCode } from "./utils";
