+++ packages/react-web3-icons/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./capitalize";
+export * from "./formatMonoSvgCode";
 export * from "./formatSymbolForIcon";
 export * from "./generateUniqueHash";
 export * from "./getAssetInfo";
